tugas 9 done

+++ main/models.py
@@ -6,7 +6,7 @@ class ChocolateProduct(models.Model):
     user = models.ForeignKey(User, on_delete=models.CASCADE)
     id = models.UUIDField(primary_key=True, default=uuid.uuid4, editable=False)
     product_name = models.CharField(max_length=255)
-    price = models.DecimalField(decimal_places=0, max_digits=5)
+    price = models.IntegerField()
     description = models.TextField()
     type = models.CharField(max_length=255)
     cocoa_ratio = models.IntegerField()
